Revert "-A_CrossZero changed -CrossZeroDataGen separated with ZeroLab -A_LF_Aprox initialized"
This reverts commit 0d035ef8
